add extended commands

+++ src/macros/MacroBuilder.ts
@@ -5,13 +5,15 @@ import * as edit from 'vscode-extension-common'
 import { Region,Lines,Modify,View,Application } from 'vscode-extension-common';
 import * as MacroRepository from './MacroRepository'
 import * as MacroExpression from './MacroExpression'
+import * as ExtendedCommands from './ExtendedCommands'
 
 const DEFAULT_SCRIPT = "// - This Feature is BETA, scripts may not work in future versions\n" +
                        "// - macro commands:\n" +
-                       "//   t - text to type\n" +
+                       "//   t - \"text to type\"\n" +
                        "//   c - cursor commands\n" +
                        "//   s - selection commands\n" +
                        "//   v - all vscode commands\n" +
+                       "//   x - macro script helper commands\n" +
                        "\n"
 
 const cursorDecorationType = View.createCursorDecoratorType()
@@ -27,14 +29,16 @@ const codeLensChangeEvent = new vscode.EventEmitter<void>()
 vscode.languages.registerCodeLensProvider({language:'plaintext' }, {
     provideCodeLenses: (document, cancel) => {
         if (document !== previewDocument) return
-        return [View.makeCodeLens('Set current text as starting point for macro', 0,0, updatePreviewTextInitialState)]
+        return [View.makeCodeLens('Set current text as starting point for macro test', 0,0, updatePreviewTextInitialState)]
     },
     onDidChangeCodeLenses: codeLensChangeEvent.event
 })
 vscode.languages.registerCodeLensProvider({language:'macro' }, {
     provideCodeLenses: (document, cancel) => {
         if (currentMacroScript.content === currentMacroScriptContentLastSave) return
-         return [View.makeCodeLens('Save Script: ' + currentMacroScript.name, 0,0, saveScript)]
+
+        const name = currentMacroScript.name ? currentMacroScript.name : ''
+        return [View.makeCodeLens('Save Script: ' + name, 0,0, saveScript)]
     },
     onDidChangeCodeLenses: codeLensChangeEvent.event
 })
@@ -42,8 +46,13 @@ vscode.languages.registerCodeLensProvider({language:'macro' }, {
 async function evalFunctionExpression(expression:string) {
     const macroExpression = MacroExpression.parseScript(expression)
     for (const command of macroExpression.commands) {
-        const parameters = [command.command].concat(command.parameters)
-        await vscode.commands.executeCommand.apply(this, parameters)
+        if (command.type === MacroExpression.CommandType.VSCODE_COMMAND) {
+            const parameters = [command.command].concat(command.parameters)
+            await vscode.commands.executeCommand.apply(this, parameters)
+        }
+        if (command.type === MacroExpression.CommandType.EXTENDED_COMMAND) {
+            ExtendedCommands.executeCommand(command.command, vscode.window.activeTextEditor, command.parameters[0])
+        }
     }   
 }
 
@@ -57,45 +66,48 @@ async function evalFunctionExpressionAsPreview(expression:string, previewEditor:
     await View.setFocusOnEditorColumn(previousColumnActive)
 }
 
+async function pickMacro():Promise<void|View.QuickPickActionable> {
+    const existingMacros = await MacroRepository.getMacroNames()
+    const options:View.QuickPickActionable[] = existingMacros.map(macroName => {
+        return View.makeOption({label:macroName, final:true, description:''})
+    })
+    const selectedItem = await View.promptOptions(options)
+    return selectedItem
+}
+
 export function runCurrentMacro() {
     evalFunctionExpression(currentMacroScript.content)
 }
 
+export async function createMacro() {
+    // TODO set default preview text with current selection if exists
+    
+    const selectedText = vscode.window.activeTextEditor.document.getText(vscode.window.activeTextEditor.selection)
+    if (selectedText) previewText = selectedText
+    else previewText = ''
+    currentMacroScript = {content:DEFAULT_SCRIPT, name:null}
+    openMacroEditor()
+}
+
 export async function runMacro() {
-    const existingMacros = await MacroRepository.getMacroNames()
-    const options:View.QuickPickActionable[] = existingMacros.map(macroName => {
-        return View.makeOption({label:macroName, final:true, description:'run macro'})
-    })
-    const selectedItem = await View.promptOptions(options)
+    const selectedItem = await pickMacro()
     if (!selectedItem) return 
 
     currentMacroScript = (await MacroRepository.fetchMacro(selectedItem.label))
     runCurrentMacro()
 }
 
-export async function createMacro() {
-    currentMacroScript = {content:DEFAULT_SCRIPT, name:null}
-    openMacroEditor()
-}
-
 export async function deleteMacro() {
-    const existingMacros = await MacroRepository.getMacroNames()
-    const options:View.QuickPickActionable[] = existingMacros.map(macroName => {
-        return View.makeOption({label:macroName, final:true, description:'delete macro'})
-    })
-    const selectedItem = await View.promptOptions(options)
+    const selectedItem = await pickMacro()
     if (!selectedItem) return 
 
-    await MacroRepository.deleteMacro(selectedItem.label)
-    return vscode.window.showInformationMessage('deleted macro: ' + selectedItem.label)
+    const answer = await vscode.window.showWarningMessage('Delete Macro: ' + selectedItem.label, 'Delete', 'Cancel')
+    if (answer === 'Delete')
+        await MacroRepository.deleteMacro(selectedItem.label)
 }
 
 export async function editMacro() {
-    const existingMacros = await MacroRepository.getMacroNames()
-    const options:View.QuickPickActionable[] = existingMacros.map(macroName => {
-        return View.makeOption({label:macroName, final:true, description:'edit macro'})
-    })
-    const selectedItem = await View.promptOptions(options)
+    const selectedItem = await pickMacro()
     if (!selectedItem) return 
 
     currentMacroScript = (await MacroRepository.fetchMacro(selectedItem.label))
@@ -108,7 +120,7 @@ async function openMacroEditor() {
     scriptDocument                  = scriptEditor.document
     const previousFocusColumnNumber = vscode.window.activeTextEditor.viewColumn
     const macroScript               = scriptEditor.document.getText()
-    let   previewEditor             = await View.openShowDocument('Macro Preview.txt', 'sourceText', false, 3)
+    let   previewEditor             = await View.openShowDocument('Macro Preview.txt', previewText, false, 3)
     previewDocument                 = previewEditor.document
     previewSelections               = previewEditor.selections
